requirements.txt
pydantic>=2.0.0
redis>=5.0.0
requests>=2.31.0
python-dotenv>=1.0.0
fastapi>=0.111.0
uvicorn>=0.30.0
structlog>=24.0.0
rich>=13.0.0
colorama>=0.4.6

# Dependencies for SNAC decoder
snac
numpy
scipy
torch --index-url https://download.pytorch.org/whl/cpu